+++ example/engine_chatbot/main.py
@@ -3,33 +3,18 @@ from typing import Dict
 import uvicorn
 from fastapi import FastAPI, Request, Response, BackgroundTasks, Query, Depends
 
-import pywce
 from example.engine_chatbot.dependencies import get_engine_instance, get_whatsapp_instance
-from pywce.engine_logger import get_engine_logger
+from pywce import Engine, pywce_logger, WhatsApp
 
-logger = get_engine_logger(__name__)
+logger = pywce_logger(__name__)
 
-# - App Metadata -
-app = FastAPI(
-    name="Pywce Bot",
-    description="An example chatbot using pywce core engine",
-    version="0.0.1",
-    contact={
-        "name": "DonnC",
-        "email": "donnclab@gmail.com",
-    },
-    license_info={"name": "MIT"},
-)
+app = FastAPI()
 
 
-# - Utility Functions -
-async def webhook_event(payload: Dict, headers: Dict, engine: pywce.PywceEngine) -> None:
+# - Utility Function -
+async def webhook_event(payload: Dict, headers: Dict, engine: Engine) -> None:
     """
-    Process webhook event in the background using pywce engine.
-
-    :param engine: Pywce engine instance
-    :param payload: WhatsApp webhook payload
-    :param headers: WhatsApp webhook headers
+    Process webhook in the background using pywce engine.
     """
     logger.debug("Received webhook event, processing..")
     await engine.process_webhook(webhook_data=payload, webhook_headers=headers)
@@ -40,15 +25,10 @@ async def webhook_event(payload: Dict, headers: Dict, engine: pywce.PywceEngine)
 async def process_webhook(
         request: Request,
         background_tasks: BackgroundTasks,
-        engine: pywce.PywceEngine = Depends(get_engine_instance)
+        engine: Engine = Depends(get_engine_instance)
 ) -> Response:
     """
     Handle incoming webhook events from WhatsApp and process them in the background.
-
-    :param request: FastAPI Request object containing the webhook payload
-    :param background_tasks: FastAPI BackgroundTasks object for async processing
-    :param engine: pywce.PywceEngine instance for webhook processing
-    :return: HTTP Response with "ACK" content
     """
     payload = await request.json()
     headers = dict(request.headers)
@@ -65,13 +45,10 @@ async def verify_webhook(
         mode: str = Query(..., alias="hub.mode"),
         token: str = Query(..., alias="hub.verify_token"),
         challenge: str = Query(..., alias="hub.challenge"),
-        whatsapp: pywce.WhatsApp = Depends(get_whatsapp_instance)
+        whatsapp: WhatsApp = Depends(get_whatsapp_instance)
 ) -> Response:
     """
-    Verify WhatsApp webhook using query parameters.
-
-    :param whatsapp: WhatsApp instance to process webhook challenge
-    :return: HTTP Response with challenge content if verification succeeds, otherwise "Forbidden"
+    Verify WhatsApp webhook callback url.
     """
     result = whatsapp.util.verify_webhook_verification_challenge(
         mode=mode, token=token, challenge=challenge
@@ -81,6 +58,7 @@ async def verify_webhook(
         return Response(content="Forbidden", status_code=403)
 
     return Response(content=result, status_code=200)
+
 
 if __name__ == "__main__":
     uvicorn.run(app, host="0.0.0.0", port=8000)