+++ CHANGELOG.md
@@ -1,16 +1,19 @@
 ## [1.0.0] - Jan 2025.
+
 * Initial release.
 
-
 ## [1.0.1] - Jan 2025.
+
 * Support for `@hook` decorator on hooks
 * Added new default session manager implementation using `cachetool` lib
 
 Decorator enables hook caching
 
 This enhance developer productivity and provides future support like hook metrics and performance
+
 ```python
 from pywce import hook, HookArg, TemplateDynamicBody
+
 
 @hook
 def username(arg: HookArg) -> HookArg:
@@ -26,3 +29,9 @@ def username(arg: HookArg) -> HookArg:
 
     return arg
 ```
+
+## [1.0.2] - Feb 2025
+
+* Added global config for toggling `read_receipts` property
+* Added global config for toggling `tag_on_reply` property to enable disable reply to message
+* CTA buttons now supported